Assign single responsible operation
