+++ createbook.rb
@@ -1,0 +1,9 @@
+def createbook
+    print "\Title: "
+    title = gets.chomp.capitalize
+    print "\Author: "
+    author = gets.chomp.capitalize
+    book = Book.new(title, author)
+    APP.addbook(book)
+    puts "\Book created successfully!"
+  end
